fixes for issues found during PR review
